Added in random number generation for the card drawing process.
Created method to check for off the blackjack win condition (starting with 21).

Also added a if statement to the placeBet method to be able to reset the current bet when going from 1 game to another, to avoid possible glitches that allow you to win the money you just lost in a bet.